+++ drift/lib/drift.dart
@@ -19,7 +19,7 @@ export 'src/runtime/executor/interceptor.dart';
 export 'src/runtime/query_builder/query_builder.dart'
     hide CaseWhenExpressionWithBase, BaseCaseWhenExpression;
 export 'src/runtime/types/converters.dart';
-export 'src/runtime/types/mapping.dart' hide BaseSqlType;
+export 'src/runtime/types/mapping.dart' hide BaseSqlType, UserDefinedSqlType;
 export 'src/utils/lazy_database.dart';
 
 /// A [ListEquality] instance used by generated drift code for the `==` and