+++ drift/lib/src/runtime/query_builder/expressions/variables.dart
@@ -8,7 +8,7 @@ part of '../query_builder.dart';
 final class Variable<T extends Object> extends Expression<T> {
   /// The Dart value that will be sent to the database
   final T? value;
-  final CustomSqlType<T>? _customType;
+  final UserDefinedSqlType<T>? _customType;
 
   // note that we keep the identity hash/equals here because each variable would
   // get its own index in sqlite and is thus different.
@@ -67,12 +67,7 @@ final class Variable<T extends Object> extends Expression<T> {
   /// database engine. For instance, a [DateTime] will me mapped to its unix
   /// timestamp.
   dynamic mapToSimpleValue(GenerationContext context) {
-    final type = _customType;
-    if (value != null && type != null) {
-      return type.mapToSqlParameter(value!);
-    } else {
-      return context.typeMapping.mapToSqlVariable(value);
-    }
+    return BaseSqlType.mapToSqlParameter<T>(context, _customType, value);
   }
 
   @override
@@ -126,7 +121,7 @@ final class Constant<T extends Object> extends Expression<T> {
   /// The value that will be converted to an sql literal.
   final T? value;
 
-  final CustomSqlType<T>? _customType;
+  final UserDefinedSqlType<T>? _customType;
 
   /// Constructs a new constant (sql literal) holding the [value].
   const Constant(this.value, [this._customType]);
@@ -142,12 +137,8 @@ final class Constant<T extends Object> extends Expression<T> {
 
   @override
   void writeInto(GenerationContext context) {
-    final type = _customType;
-    if (value != null && type != null) {
-      context.buffer.write(type.mapToSqlLiteral(value!));
-    } else {
-      context.buffer.write(context.typeMapping.mapToSqlLiteral(value));
-    }
+    return context.buffer
+        .write(BaseSqlType.mapToSqlLiteral(context, _customType, value));
   }
 
   @override