+++ drift/test/database/types/custom_type_test.dart
@@ -11,30 +11,33 @@ void main() {
 
   group('in expression', () {
     test('variable', () {
-      final c = Variable<UuidValue>(uuid, const UuidType());
+      final c = Variable<UuidValue>(uuid, const NativeUuidType());
 
-      expect(c.driftSqlType, isA<UuidType>());
+      expect(c.driftSqlType, isA<NativeUuidType>());
       expect(c, generates('?', [uuid]));
     });
 
     test('constant', () {
-      final c = Constant<UuidValue>(uuid, const UuidType());
+      final c = Constant<UuidValue>(uuid, const NativeUuidType());
 
-      expect(c.driftSqlType, isA<UuidType>());
+      expect(c.driftSqlType, isA<NativeUuidType>());
       expect(c, generates("'$uuid'"));
     });
 
     test('cast', () {
-      final cast = Variable('foo').cast<UuidValue>(const UuidType());
+      final cast = Variable('foo').cast<UuidValue>(const NativeUuidType());
 
-      expect(cast.driftSqlType, isA<UuidType>());
+      expect(cast.driftSqlType, isA<NativeUuidType>());
       expect(cast, generates('CAST(? AS uuid)', ['foo']));
     });
   });
 
   test('for inserts', () async {
-    final executor = MockExecutor();
-    final database = TodoDb(executor);
+    final sqlite3Executor = MockExecutor();
+    final postgresExecutor = MockExecutor();
+    when(postgresExecutor.dialect).thenReturn(SqlDialect.postgres);
+
+    var database = TodoDb(sqlite3Executor);
     addTearDown(database.close);
 
     final uuid = Uuid().v4obj();
@@ -42,23 +45,48 @@ void main() {
         .into(database.withCustomType)
         .insert(WithCustomTypeCompanion.insert(id: uuid));
 
-    verify(executor
-        .runInsert('INSERT INTO "with_custom_type" ("id") VALUES (?)', [uuid]));
+    verify(sqlite3Executor.runInsert(
+        'INSERT INTO "with_custom_type" ("id") VALUES (?)', [uuid.toString()]));
+
+    database.close();
+    database = TodoDb(postgresExecutor);
+
+    await database
+        .into(database.withCustomType)
+        .insert(WithCustomTypeCompanion.insert(id: uuid));
+
+    verify(postgresExecutor.runInsert(
+        r'INSERT INTO "with_custom_type" ("id") VALUES ($1)', [uuid]));
   });
 
   test('for selects', () async {
-    final executor = MockExecutor();
-    final database = TodoDb(executor);
-    addTearDown(database.close);
-
     final uuid = Uuid().v4obj();
-    when(executor.runSelect(any, any)).thenAnswer((_) {
+
+    final sqlite3Executor = MockExecutor();
+    when(sqlite3Executor.runSelect(any, any)).thenAnswer((_) {
+      return Future.value([
+        {'id': uuid.toString()}
+      ]);
+    });
+
+    final postgresExecutor = MockExecutor();
+    when(postgresExecutor.dialect).thenReturn(SqlDialect.postgres);
+    when(postgresExecutor.runSelect(any, any)).thenAnswer((_) {
       return Future.value([
         {'id': uuid}
       ]);
     });
 
+    var database = TodoDb(sqlite3Executor);
+    addTearDown(database.close);
+
     final row = await database.withCustomType.all().getSingle();
     expect(row.id, uuid);
+
+    await database.close();
+    database = TodoDb(postgresExecutor);
+
+    final pgRow = await database.withCustomType.all().getSingle();
+    expect(pgRow.id, uuid);
   });
 }